% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimates.R
\docType{data}
\name{estimates}
\alias{estimates}
\title{Estimates}
\format{
A tibble with 1,026 rows and
7 variables:

\describe{
\item{\strong{Community}}{\tabular{ll}{
   \emph{Type:} \tab factor \cr
    \tab  \cr
   \emph{Description:} \tab Counties \cr
    \tab  \cr
   \emph{Levels:} \tab Albany, Allegany, Broome, Cattaraugus, Cayuga, Chautauqua, Chemung, Chenango, Clinton, Columbia, Cortland, Delaware, Dutchess, Erie, Essex, Franklin, Fulton, Genesee, Greene, Hamilton, Herkimer, Jefferson, Lewis, Livingston, Madison, Monroe, Montgomery, Nassau, Niagara, Oneida, Onondaga, Ontario, Orange, Orleans, Oswego, Otsego, Putnam, Rensselaer, Rockland, Saratoga, Schenectady, Schoharie, Schuyler, Seneca, St. Lawrence, Steuben, Suffolk, Sullivan, Tioga, Tompkins, Ulster, Warren, Washington, Wayne, Westchester, Wyoming, Yates \cr
}


}
\item{\strong{Year}}{\tabular{ll}{
   \emph{Type:} \tab factor \cr
    \tab  \cr
   \emph{Description:} \tab Period \cr
    \tab  \cr
   \emph{Levels:} \tab 2017, 2018, 2019 \cr
}


}
\item{\strong{Age}}{\tabular{ll}{
   \emph{Type:} \tab factor \cr
    \tab  \cr
   \emph{Description:} \tab Age stratification \cr
    \tab  \cr
   \emph{Levels:} \tab 18-34, 35-54, 55-64 \cr
}


}

\item{\strong{Sex}}{\tabular{ll}{
   \emph{Type:} \tab factor \cr
    \tab  \cr
   \emph{Description:} \tab Sex stratification \cr
    \tab  \cr
   \emph{Levels:} \tab Female, Male \cr
}


}

\item{\strong{Population}}{\tabular{ll}{
   \emph{Type:} \tab double \cr
    \tab  \cr
   \emph{Description:} \tab Population \cr
    \tab  \cr
}


}

\item{\strong{Proportion of OUD Estimation}}{\tabular{ll}{
   \emph{Type:} \tab double \cr
    \tab  \cr
   \emph{Description:} \tab Proportion of OUD Estimation \cr
    \tab  \cr
}


}

\item{\strong{Proportion of Undiagnosed OUD Estimation}}{\tabular{ll}{
   \emph{Type:} \tab double \cr
    \tab  \cr
   \emph{Description:} \tab *Proportion of Undiagnosed OUD Estimation \cr
    \tab  \cr
}


}
}
}
\usage{
data(estimates)
}
\description{
This data set contains the model implied OUD prevalence and
un-diagnosed OUD proportions
}
\keyword{datasets}
